The big utils/ cleanup. Among other things, this replaces the Base64, MD5
and SHA1 code with faster implementations. Most of the APIs affected are
seldom used in the codebase, hence the name of the branch.

- utils/base64, md5, sha1: use faster, cleaner implementations
  [873283ec] [468b88e5] [6ea69afb]
- utils/strhex, rc4, crc8, crc32: accept (void *) and size_t as arguments
  [6a96ba82] [5a5299ed] [5cdbfe73] [47937fce]
- luaapi/lib: move Lua bindings into lualib/
  [bfce4543]
- lualib/json: clean up code, add unit test coverage
  [b07edb06] [bc13ce2a]
- remove checks for CPU endianness
  [dbad430b] [12ec60f1]
- add reader and writer functions for 32-bit integers
  [71a98724] [17fd6b9c]

+++ src/astra/core/spawn.c
@@ -197,7 +197,7 @@ int fork_and_exec(const char *command, pid_t *out_pid
         fcntl(STDERR_FILENO, F_SETFD, 0);
 
         /* reset signal handlers and masks */
-        for (size_t sig = 1; sig < NSIG; sig++)
+        for (int sig = 1; sig < NSIG; sig++)
         {
             struct sigaction sa;
             if (sigaction(sig, NULL, &sa) == 0 && sa.sa_handler != SIG_DFL)
